+++ src/src/Personajes/NPC.h
@@ -8,7 +8,7 @@ class Motor;
 
 class NPC : public Character {
 public:
-	NPC(short _i_id, float _i_x, float _i_y, float _i_z, short _i_vida, short _i_velocidad, short _i_danyo_ataque_normal, 
+	NPC(short _i_id, float _i_x, float _i_y, float _i_z, short _i_vida, float _i_velocidad, short _i_danyo_ataque_normal, 
 	short _i_danyo_ataque_fuerte, Enum_Tipo_Enemigo _i_tipo_npc);
 	
 	virtual ~NPC();
