+++ src/src/Personajes/NPC/Beast.h
@@ -5,7 +5,7 @@
 
 class Beast : public Enemy {
 public:
-	Beast(short _i_id, float _i_x, float _i_y, float _i_z, short _i_vida, short _i_velocidad, short _i_danyo_ataque_normal, 
+	Beast(short _i_id, float _i_x, float _i_y, float _i_z, short _i_vida, float _i_velocidad, short _i_danyo_ataque_normal, 
 	short _i_danyo_ataque_fuerte);
 	
 	virtual ~Beast();
