+++ src/src/Personajes/NPC/Enemy_Normal.h
@@ -5,7 +5,7 @@
 
 class Enemy_Normal : public Enemy {
 public:
-	Enemy_Normal(short _i_id, float _i_x, float _i_y, float _i_z, short _i_vida, short _i_velocidad, short _i_danyo_ataque_normal, 
+	Enemy_Normal(short _i_id, float _i_x, float _i_y, float _i_z, short _i_vida, float _i_velocidad, short _i_danyo_ataque_normal, 
 	short _i_danyo_ataque_fuerte);
 
 	virtual ~Enemy_Normal();
